+++ app/layout.tsx
@@ -1,10 +1,10 @@
-import Footer from "@/app/Footer";
-import Header from "@/app/Header";
-import { Providers } from "@/app/providers";
+import { Providers } from "@/components/providers/providers";
 import { Metadata } from "next";
 import { Inter } from "next/font/google";
 
 import "./globals.css";
+import Header from "@/components/layout/Header";
+import Footer from "@/components/layout/Footer";
 
 const inter = Inter({ subsets: ["latin"] });
