+++ components/layout/Footer.tsx
@@ -1,9 +1,25 @@
+import Image from 'next/image';
+
 const Footer = () => {
-  return (
-    <footer className="bg-[#020F3B] flex justify-center w-full p-2 text-xl text-slate-600">
-      <p>©LikeHome by Stellar Horizons. CMPE-165 Fall 2024.</p>
-    </footer>
-  );
+    return (
+    <footer className="bg-customBlue py-[50px] lg:py-[50px]">
+        <div className="container mx-auto">
+                <div className="flex flex-col md:flex-row md:justify-between gap-6">
+                    {/* logo */}
+                    <div className="flex items-center gap-5 justify-center xl:w-max">
+                        <Image
+                            src="/stellarHorizons.png"
+                            alt="LikeHome"
+                            width={55}
+                            height={20}
+                        />
+                        <span className="ml-2 text-lg font-bold text-white">LikeHome</span> {/* Text displayed next to image */}
+                    </div>
+                </div>
+        </div>
+        </footer>
+    );
+
 };
 
 export default Footer;
