+++ components/layout/Header.tsx
@@ -1,4 +1,5 @@
 import Image from "next/image";
+import ThemeSwitch from "./ThemeSwitch";
 
 const Header = () => {
   return (
@@ -13,32 +14,37 @@ const Header = () => {
               width={55}
               height={20}
             />
-            <span className="ml-2 text-lg font-bold ">LikeHome</span>
+            <span className="ml-2 text-lg font-bold">LikeHome</span>{" "}
             {/* Text displayed next to image */}
           </div>
-          {/* sign in & register */}
-          <div className="flex items-center gap-4">
+          <div className="flex items-center gap-8 ">
             {/* Navigation Links */}
             <nav className="flex items-center gap-8">
+              {" "}
               {/* Flex container for navigation items */}
-              <div className="text-gray-700 hover:text-purple-600 cursor-pointer">
+              <div className="text-black-700 hover:text-purple-600 cursor-pointer">
                 Home
               </div>
-              <div className="text-gray-700 hover:text-purple-600 cursor-pointer">
+              <div className="text-black-700 hover:text-purple-600 cursor-pointer">
                 Rooms
               </div>
-              <div className="text-gray-700 hover:text-purple-600 cursor-pointer">
+              <div className="text-black-700 hover:text-purple-600 cursor-pointer">
                 Contact
               </div>
             </nav>
+            {/* sign in & register */}
+
             <button className="px-4 py-2 bg-black text-white rounded hover:bg-black-500">
+              {" "}
               {/* Sign In Button */}
               SIGN IN
             </button>
-            <button className="px-4 py-2 bg-purple-600 text-white rounded hover:bg-purple-600">
+            <button className="px-4 py-2 bg-purple-600 text-white rounded hover:bg-purple-700">
+              {" "}
               {/* Register Button */}
               REGISTER
             </button>
+            <ThemeSwitch />
           </div>
         </div>
       </div>
