+++ components/providers/providers.tsx
@@ -1,4 +1,3 @@
-
 import { ThemeProvider } from 'next-themes';
 
 export const Providers = ({ children }: { children: React.ReactNode }) => {
